requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
flet==0.25.2
PyPDF2==3.0.1
python-docx==1.1.2
reportlab==4.2.5
PyJWT==2.10.1

# Email automation
aiosmtplib==3.0.1
email-validator==2.1.0
jinja2==3.1.6

# Web search
tavily-python==0.3.4

# Async utilities
aiofiles==24.1.0
nltk==3.8.1

# Supabase official client (optional but recommended)
supabase==2.4.5

# Phase 3.2: Advanced RAG with LangChain
langchain==1.2.15
sentence-transformers==3.0.0
faiss-cpu==1.13.2
numpy==1.26.4
